Implement smooth toast transitions with queue system and build fixes
Major improvements:
- Replace single pending toast with queue system for better rapid key press handling
- Ensure smooth fade-out → fade-in transitions even with rapid toast requests
- Adjust lyrics widget position to 50px from top of screen
- Add -Wall -Werror compilation flags for stricter code quality

Technical changes:
- Replace std::optional<PendingToast> with std::queue<PendingToast> in Player class
- Add MAX_TOAST_QUEUE_SIZE constant (10) to prevent memory issues
- Update showToast() method to queue multiple toasts instead of replacing
- Update updateGUI() toast management to process queue sequentially
- Fix numerous signed/unsigned comparison warnings across codebase
- Fix member initialization order warnings in Audio constructor
- Fix unused function warnings with __attribute__((unused))

Files modified:
- include/player.h: Updated toast system architecture
- src/player.cpp: Implemented queue-based toast management
- src/LyricsWidget.cpp: Adjusted position from 20px to 50px from top
- src/Makefile.am: Added -Wall -Werror compilation flags
- src/audio.cpp, src/fft.cpp, src/playlist.cpp: Fixed signed/unsigned warnings
- src/ChainedStream.cpp: Fixed comparison type warnings
- src/system.cpp: Marked unused function with attribute

+++ src/ChainedStream.cpp
@@ -63,7 +63,7 @@ ChainedStream::ChainedStream(std::vector<TagLib::String> paths)
                 first_channels = temp_stream->getChannels();
             } else {
                 // Ensure all subsequent tracks have the same format.
-                if (temp_stream->getRate() != first_rate || temp_stream->getChannels() != first_channels) {
+                if (temp_stream->getRate() != static_cast<unsigned int>(first_rate) || temp_stream->getChannels() != static_cast<unsigned int>(first_channels)) {
                     throw InvalidMediaException("ChainedStream tracks must have the same sample rate and channel count.");
                 }
             }
